Jepsen client can now hit the /ping marathon endpoint
Summary:
Now the jepsen client is able to hit the marathon api `/ping` endpoint.
Will work on catching the pong reply in the next PR.
Also, I got rid of all the hardcoded localhost addresses everywhere and made it more generic.

Test Plan: Check by going inside the VM if mesos, marathon and zookeeper are up and running and if by hitting the /ping we get a pong reply.

Reviewers: ichernetsky, jenkins

Reviewed By: ichernetsky, jenkins

Subscribers: marathon-dev

Differential Revision: https://phabricator.mesosphere.com/D890