Use event timestamp in GracefulTaskKillIntegrationTest. (#5430)
Summary:
The tests fails sometimes with "4160 was not less than 4000". Since we
take the receive time stamp after it was processed by the test we might
have some delay. To avoid the extra cost we'll take the timestamp when
the event was received by Marathon.

Test Plan: sbt integration:test

Reviewers: timcharper, unterstein, jenkins, ichernetsky

Reviewed By: timcharper, jenkins, ichernetsky

Subscribers: marathon-team, marathon-dev, jenkins

JIRA Issues: MARATHON-7468

Differential Revision: https://phabricator.mesosphere.com/D832