Jepsen code for spinning up a zookeeper cluster and running Marathon in HA mode.
Summary: Now Marathon runs in High availability mode. Also, now a unified zookeeper cluster spins up where there is one leader and rest are the followers. The previous version of this code was setting up multiple cluster where each node was a self contained and no 2 nodes were communicating with each other which will not react to network splits at all. Now the setup is like the production cluster setup. In the next PR, I will be changing the code for running Marathon Framework and will use systemd instead.

Test Plan: Verified if the entire setup is up and running by going inside the VMs and checking if the setup is correct and manually checking the status of the running apps.

Reviewers: ichernetsky, kensipe, jenkins

Reviewed By: ichernetsky, jenkins

Subscribers: marathon-dev

Differential Revision: https://phabricator.mesosphere.com/D910