Jepsen checker code, db code to pull log files and bug fixes.
Summary:
Added Jepsen checker code which checks if the apps acknowledged by Marathon are healthy after the network heals for good after constant splits and heals cycle generated via Jepsen.
Additionally, added code to the `db` module to pull logs for Mesos, Marathon and Zookeeper which is downloaded just before the teardown phase and stored in the `store` directory.
Made a log of bug fixes and made sure that the necessary exceptions are caught.

Test Plan: Check for a Jepsen cycle which passes and one which fails. Also if all teardown is cleaning the VMs completely.

Reviewers: ichernetsky, kensipe, jenkins

Reviewed By: ichernetsky, kensipe

Subscribers: marathon-dev

Differential Revision: https://phabricator.mesosphere.com/D931