Jepsen test for updating apps and then re-updating apps with intermediate network failures.
Summary:
This jepsen test basically creates and deploys apps for 100 seconds, then updates few of those apps for another 30 seconds by increasing the number of instances to 1 more than the number of available nodes. The update also adds a constraint of deploying each task instance of the app on unique host. Further, some apps are again updated by decreasing the instances to the number of hosts with the same constraint which should allow the app to get deployed. The checker code will be added in the next PR once this test spec is reviewed and verified.
Also, I moved around some functions to reorganize the code a bit.

Test Plan: Run the test and verify the behaviour with the downloaded logs.

Reviewers: kensipe, ichernetsky, timcharper, jenkins

Reviewed By: ichernetsky, jenkins

Subscribers: marathon-dev

Differential Revision: https://phabricator.mesosphere.com/D951