Integrated marathon-cli tool aka mcli tool into Jepsen testing.
Summary: `mcli` tool is a jq powered mesos cli tool, written by Tim Harper, which has a `marathon-cli` tool as well. This tool reads and interprets Marathon HTTP responses. This tool is used in Jepsen testing to create neat logs for apps and tasks without the need of going into the marathon logs and making continuous curl requests as this tool does this for you.

Test Plan: verify if the all the logs are updated to date and retrieved, at the end of the Jepsen test.

Reviewers: ichernetsky, kensipe, timcharper, jenkins

Reviewed By: ichernetsky, jenkins

Subscribers: marathon-dev

Differential Revision: https://phabricator.mesosphere.com/D964