Jepsen scale test to randomly scale up and scale down apps.
Summary: Added a new test which basically randomly picks a scale up or scale down operation on a random app. Adding to this, the test also creates network partitions after performing one of the scale up or scale down operations. Also, this PR fixes some bugs which now makes the checking the status of the apps significantly faster.

Test Plan: Verify with the logs if the test scales up and scales down the apps as expected.

Reviewers: ichernetsky, kensipe, jenkins

Reviewed By: ichernetsky, kensipe, jenkins

Subscribers: marathon-dev

Differential Revision: https://phabricator.mesosphere.com/D967