Added `agentId` to the pod status (#5448)
Summary: To align the pod api with the app api, the `agentId` property for running instances is introduced. Furthermore the agentId need for different purposes (e.g. accessing the mesos sandbox for a given container), so it makes sense to expose the agentId here.

Test Plan: sbt test

Reviewers: jdef, jeschkies, zen-dog, jenkins, timcharper

Reviewed By: jdef, jenkins, timcharper

Subscribers: tillt, marathon-dev, marathon-team

JIRA Issues: MARATHON-7628

Differential Revision: https://phabricator.mesosphere.com/D926