Destroy app Jepsen test and middleware to catch all the exceptions.
Summary: This PR adds a new test which randomly destroys an app and introduces network splits as well. Also, this PR introduces a middleware to catch all the exceptions for all the http requests for all the tests. The advantage of doing this is that there are a lot of repeated exceptions to catch and this was causing to adding a lot of redundant code for each test. Now, all the status code exceptions are caught at one place. Basically this a wrapper on the http client.

Test Plan: Make sure all the tests run as expected and all the exceptions are caught by the middleware.

Reviewers: ichernetsky, kensipe, jenkins

Reviewed By: ichernetsky, kensipe

Subscribers: marathon-dev

Differential Revision: https://phabricator.mesosphere.com/D971