Reset root group cache when elected as leader (#5461)
* Reset root group cache when elected as leader

Summary:
During the replacement of GroupManagerActor to GroupManagerImpl, the following behavior was introduced:
When taking over leadership, a marathon instance will not load current root group from zookeeper, instead the old (probably outdated) root group will used to operate. This leads wrong behavior, because this root group is not in sync with the current zookeeper state.

This fix introduced a refresh cache implementation which is called during starting after taking leadership.

Test Plan: sbt integration:test

Reviewers: zen-dog, timcharper, jeschkies, jenkins

Reviewed By: zen-dog, jeschkies, jenkins

Subscribers: marathon-dev, marathon-team

JIRA Issues: MARATHON-7458

Differential Revision: https://phabricator.mesosphere.com/D847

* Fix caching issues after performing migration or restoring a backup

IntegrationTest current in progress

Summary:
GroupManager and GroupRepository are holding in memory caches of the root group. The cache is loaded when it is accessed the first time.
Actually this is really bad, because each marathon will log the amount of groups during startup through Kamon.
Therefore the root group state is loaded from zk when the marathon instance is started.
When the marathon instance is elected as leader, this cache is still in the same state as the time marathon started.
Therefore we need to re-load the root group from zk again from zookeeper when becoming leader.
The same is true after doing the migration. A migration or a restore also affects the state of zookeeper, but does not
update the internal hold caches. Therefore we need to refresh the internally loaded caches after the migration.
Actually we need to do the fresh twice, before the migration, to perform the migration on the current zk state and after
the migration to have marathon loaded the current valid state to the internal caches.

Test Plan: sbt integration:test-only *BackupRestoreIntegrationTest

Reviewers: jeschkies, timcharper, zen-dog, jenkins

Reviewed By: jeschkies, zen-dog, jenkins

Subscribers: jdef, marathon-dev, marathon-team

JIRA Issues: MARATHON-7565

Differential Revision: https://phabricator.mesosphere.com/D895

* adapted implementation to 1.4

* docs

* feedback