Fix flaky `TaskKillingIntegrationTest`
Summary:
The test waits for a `TASK_KILLING` event however in rare cases the task will fail shortly after start.
Killin a failed task results in `TASK_UNKNOWN` hence this test will never receive the event it's waiting for.
Extended the test to handle this edge case.

Test Plan: it

Reviewers: jeschkies, ichernetsky, timcharper, jenkins

Reviewed By: jeschkies, ichernetsky, jenkins

Subscribers: jenkins, marathon-dev, marathon-team

Differential Revision: https://phabricator.mesosphere.com/D985