Update AgentInfo when launching an instance based on an existing reservation.
Summary:
When an agent (e.g. "agent-1") reboots, the agentId might change under certain circumstances (to, e.g. "agent-2"). If that agent's state contained dynamic reservations and persistent volumes, Marathon would not update the AgentInfo when launching a new instance based on these. As a result, the instance as reported via the Marathon API would still say "agent-1", when the actual task and the persistent volume would run on "agent-2"; the volume and task appear to be on different nodes.

This commit will update the AgentInfo for each LaunchOnReservation.

Test Plan: sbt test

Reviewers: jeschkies, zen-dog, kensipe, jenkins, jdef, meichstedt

Reviewed By: jeschkies, zen-dog, jenkins

Subscribers: jdef, marathon-dev, marathon-team

JIRA Issues: MARATHON-7707

Differential Revision: https://phabricator.mesosphere.com/D1002