Updated README and added usage of environment variable in Vagrantfile.
Summary: This PR deals with updating the outdated README file with the correct usage. Also, now the vagrantfile picks up an environment variable if set, to spin up more vms if the user wants to. Finally, the Makefile was updated accordingly to update the usage.

Test Plan: Check if the setting of environment variable works with vagrant and test the updated usage.

Reviewers: ichernetsky, kensipe, jenkins

Reviewed By: ichernetsky, kensipe, jenkins

Subscribers: marathon-dev

Differential Revision: https://phabricator.mesosphere.com/D1012