Distinguish app names in GracefulTaskKillIntegrationTest.
Summary:
The test sometimes fails with

`4153 was not less than 4000 the task kill event took longer than the task kill grace period`

So far we think this is just some flakiness. This change introduces
proper names so we can distinguish the apps in the archived sandbox
logs. This way we can verify that one was killed with SIGTERM and the
over with SIGKILL

Test Plan: pipeline

Reviewers: unterstein, ichernetsky, jenkins

Reviewed By: unterstein, ichernetsky, jenkins

Subscribers: marathon-team, marathon-dev, jenkins

Differential Revision: https://phabricator.mesosphere.com/D1017