Enable successful SI Test runs with cluster of agents with different core counts (#5581)
The test was originally written assuming an agent core count of 4 which
has been true for years. Based on cost cutting measures the latest
clusters are made of 2 core agents. However CCM still creates the 4 core
count agents. This change allows this test to pass for changes in the agent
core count.