Simplify LeaderElection by using an event based approach (#5592)
* Simplify LeaderElection by using an event based approach

We encapsulate Curator election logic using a generic Akka Stream source of
events: ElectedLeader, Standby(currentLeader). Previously, we used to query
ZooKeeper each time an http endpoint was hit on the non leader. Now, we use
Zookeeper watches to be notified of leadership changes.

By using a generic stream interface, PsuedoElectionService becomes an
incredibly simple stream that simply emits a leader elected event when
it starts, and ends with a standby event when closed.

Testing also becomes easier with event production and event reacting
concerns better separated.

Further, we remove all uses of locks in ElectionService. These were
previously a source of deadlock, and a source of unseen complexity.

Finally, we update to Curator 4.x.x to overcome CURATOR-358, which induced a source of indeterminism (which manifested primarily in the unit tests).