Fix for event stream tests with multimaster (#5795)
Summary:
Event stream tests were failing after switch to multimaster. The reason for that is that all the commands were talking to (mesos) master but there are multiple masters now so the commands might be executed on different nodes. Now I do the curl always on marathon leader to make that deterministic.

JIRA issues: MARATHON-7943