Remove app and pod transitive index from groups.  (#5805)
Summary:
This is another speed up for groups. Currently we build an index over
transitive apps and pods which is costly. This index is not required
since we have an index over all transitive groups and each group has an
index of its apps and pods.

Depends on #5804.

JIRA issues: MARATHON-7945