Replace transitive apps and pods set with an iterable.  (#5807)
Summary:
This change builds on top of #5805 and cuts the execution time
of the assign dynamic service ports.

The current master takes ~18ms for 15 groups and 10,000 apps
in the dynamic port assignment benchmark. This change speeds
it up to <5ms.

I also added a benchmark for building the root group which implies
a lot of `updateApp` calls. Here, the current master takes >97000ms
for building a root group with 20 nested child groups and 1000 apps
each, i.e. building a group with 20,000 apps. This patch speeds up
the building time by two magnitudes to <230ms.

The app, pod and child group access time stays constant while the
master as a linear time in terms of number of groups.

The speed is achieved by removing the build up of an apps and pods
index (transitiveAppsById and transitivePodsById). The look up
stays constant by leveraging the constant lookup of child groups.

JIRA issues: MARATHON-7946