Benchmark and speed up service port assignment by up to three magnitudes. (#5787) (#5816)
Summary:
As @zen-dog found a big performance degradation when Marathon has already 5000 apps and a new app is added. He found that `GroupManagerImpl.assignDynamicServicePorts` had an exponential runtime.

The benchmark and flamegraph showed that most time was spend in copying and building the `RootGroup` after the update. It also seems that we update all apps even when the get the same port assigned again. This change introduces a filter so that only apps that are new or changed are updated in the `RootGroup`.

You can run the benchmark with
```
sbt "benchmark/jmh:run -i 1 -wi 3 -f1 -t1 .*GroupManagerBenchmark"
```

This is only a fix for small app updates. If many apps are updated at once we will still hit this bottleneck. We should look into the creation of `transitiveAppsById` and `transitivePodsById`
which take up most of the runtime.

JIRA issues: MARATHON_EE-1770, MARATHON_EE-1764