Transfers bytes in proxy and flush after each write.
Summary:
The proxy was broken for the event stream. Our copy method is blocking
as is `com.google.common.io.ByteStreams.copy`. Thus we would never flush
the output stream and the client would not receive any data. This patch
reintroduces the old transfer method with flushes after each write.

This diff should be followed up with an asynchronous copying so that no
threads are blocked.

JIRA issues: MARATHON-7919