Provision node and run SI tests with pipenv.  (#5831)
Summary:
This introduces pipenv to manage our system integration test dependencies
and environment. This way we can introduces and remove dependencies
without changing the Jenkins Docker images and maintain reproducible runs.

An environment is initialized with make init.
The tests are run with with make test.

You can also jump into the shell of the environment with pipenv shell.