Do not add SELF SANDBOX_PATH volumes to a pod's executor (#5870)
The code which was adding this kind of volumes to pod executors
was unreachable up until recently. After it got fixed, it was working
fine with older versions of Mesos, but in September, 2017,
the validation logic was made stricter, requiring an absolute path
for container_path in case of SELF SANDBOX_PATH volumes. But it turned
out that there is no need at all to add this kind of volumes to
pod executors in order to make ephemeral volumes work. Therefore,
delete the code completely.

JIRA issue: https://jira.mesosphere.com/browse/MARATHON-8004