Create the migration flag only while migrations are performed
Currently, the flag is created unconditionally for the duration of
checking the current version in ZK and performing migrations if necessary.
There is no need to create the flag for the duration of checking
the current version, it should be created only if and when there is
a need to perform migrations.

JIRA issues:
https://jira.mesosphere.com/browse/MARATHON-8044