Serialize full and light events only once. (#5933)
Summary:
Currently each SSE handle is serializing the events. This
is a small overhead. It should be enough to serialize them
once and then broadcast to each handle actor.

The next step is to replace the HttpEventStreamHandleActor
with an Akka stream so that we do not have to deal with
slow consumers.

JIRA issues: MARATHON-8035