Decouple storage version from build version.
Summary:
Currently the storage version we want to migrate to is defined by the
build version. This makes it hard to ship migrations with patches.
Furthermore the new versioning defined in MARATHON_EE-1759 could brake
migrations. Thus this patch decouples the version of a migration from
the build version.

The target storage version is now inferred from the steps of a
migration.

JIRA issues: MARATHON_EE-1845