Complete deployment after it was removed to avoid race condition. (#5944)
Summary:
The `DeploymentActor` is sending a `DeploymentFinishd` messaged to the
`DeploymentManagerActor` *and* fulfills a promis for the
`MarathonSchedulerActor` which triggers events. This can lead to a race
condition. The scheduler might anounce a finished deployment while it's
still in the queue of active deployments. This change has the manager
fulfill the promise *after* the deployment was removed from the queue.
Thus we avoid the race condition.

Backported from #5939 

JIRA issues: MARATHON-8039