Move flaky universe integration test to regression test only covering past problem (#5985)
Summary:
The system integration test to install the Neo4j universe package is flaky some times. For whatever reason it is pretty stable installing it manually, maybe it takes too long time to spin up and start properly. For this reason this PR will remove this system integration test and add a test which tests only the regression marathon had in March 2017. The commit fixed this behavior for the Marathon 1.4.1 release is 7e8cfa4. The initial problem was, that resident tasks with user network never became healthy. This new integration test only reflects this regression and keeps the other moving parts out of the test :)

There is a ticket to track efforts to refactor the Neo4j universe package here https://jira.mesosphere.com/browse/DCOS_OSS-2126