Version Marathon and name its artifacts. (#5970)
Summary:
With this change Marathon will be version with <major>.<minor>.<build-no>.
The build number is the number of commits since the last branch point of
a release branch. E.g. `releases/1.5` was branched at `aa89e89`. This is the
common ancestor for 1.5 and the current release. If master or `releases/1.6`
is 42 commits ahead of `aa89e89`, the build number is 42.

The build artifacts are named as follows
* `marathon-1.6.42-3e42d0d.tgz` for Tarball
* `marathon-docs-1.6.42-3e42d0d.tgz` for dcouments Tarball
* `marathon_2.12-1.6.42.jar` for Jar
* `marathon_1.6.42_all.deb` for Debian
* `marathon-1.6.42.3e42d0d-1.noarch.rpm` for RPM
* `mesosphere/marathon:v1.6.42` for Docker images

That means the tarball packages include the commit hash in their names
so that it is easier to find the commit when we integrate a build into
DC/OS.

JIRA issues: MARATHON_EE-1759