Do not send contradicting deployment events for cancellation. (#6018)
Summary:
When a deployment is cancelled by the `MarathonSchedulerActor` it will
trigger a `deployment_success` event. However, the
`DelploymentManagerActor` will trigger a `deployment_failed` event for
the same deployment. This change removes the first event so that a user
will recieve only one event. Ideally we would send a
`deployment_cancelled` event instead. That would break the API, though.