Pass instance id to offer matcher.
Summary:
Our mid term goal is to generate the instance id in the business logic
and push it towards the task launch layer. This is the first step to
generate the id in the task launcher instead of the offer matcher.

The next step should pass the id to the task launcher instead of
generating it there.