Add support for symlinks for `--plugin_dir` and `--plugin_conf` parameters
by internally using `getCanonicalFile` method on the File object. `getCanonicalFile` first converts this pathname
to absolute form if necessary, as if by invoking the getAbsolutePath() method, and then maps it to its unique form
in a system-dependent way. Among other things this typically involves resolving symbolic links (on UNIX platforms).