Test Marathon upgrades 1.4.9 thru 1.6.322. (#6113)
Summary:
Sometimes we are unsure if a Marathon upgrade would remove tasks.
E.g. with the task id changes in #6076 we require tests that upgrade
Marathon. This change introduces such a test infrastructure.

The test download Marathon 1.4.9, 1.5.6 and 1.6.322 and deploys an
app. It then starts on Marathon after the other and verifies that the app
is still running in the new version. I will add different types of apps
in a future release.

JIRA issues: MARATHON-8141