Rework Mesos v1 client
We remove the mutable connection info and frameworkId, and instead return a Source[MesosClient, NotUsed], consistent with Akka HTTP response handling. The MesosClient is yielded after all of this data is available, and contains it immutably. To do this, the Mesos graph pulls out the initial Subscribed event, and then passes the rest of the stream for further consumption
via a materializable-once source in mesosClient.mesosSource.