Introduce `Provisoned` and `Scheduled` condition.
Summary:
We would like to create instances in a scheduled state *before* they are
launched. That is why we introduce a `Scheduled` and `Provisioned`
condition.

So far instances require agent info. However, when an instance is just
scheduled there is no meanningful agent info. So it has to be optional.
All requests that require this info filter out instances that have no
info. This way we stay backwards compatible since there are no instance
in scheduled condition currently it is safe to ignore them.

JIRA issues: MARATHON-8168