Async deployment responses (#6187)
In this change we convert deployment creating API methods to use Jersey 2 AsyncResponse. By doing this, we no longer timeout ("Futures timed out"); if a deployment takes 2 minutes to schedule, and the HTTP client does not have a shorter receive timeout configured, then the response will be delivered after 2 minutes. This is decidedly better than timing out server-side, responding as such, and then doing the work anyways.

Also, we consolidate some response logic, converging places where there where multiple response generators to just simply prefer the MarathonExceptionMapper instead.

Further, we introduce Rejection, which models unknown reasons why an API request will not be satisfied, and carries with it a guarantee that the exception mapper will know how to render a response for it.

JIRA Issues: MARATHON-8174