Report successful and failed loop builds. (#6209)
Summary:
We missed a commit that caused flaky tests because the Marathon loop
itself was not working. This change introduces a count metric in
DataDog. We will set up a monitor that alarms if the number of
successful builds is below a certain threashold in the last 24hrs. In
the future we might want to add commit events so that we can triage the
problematic commits faster.

JIRA issues: MARATHON-8199