Switch to Scalatest JUnitXMLReporter to regain the timestamp field (#6225)
Back port of #6214

SBT 1.x ships with a JUnit XML reporter which is enabled, by
default. Unfortunately, this reporter is missing a field which is
expected by Jenkins: timestamp. We use this field in our test analysis
tool to reason about which tests ran in parallel.

Switching to the Scalatest JUnitXML reporter brings the field back for
the time being.

See https://github.com/sbt/sbt/issues/4153