Move integration tests to its own project
In this commit we simplify the configuration of integration tests and unit tests by moving integration tests to its own project, rather than separating the two with tags. This is arguably a less hacky-solution to the problem of having different configurations for running different kinds of tests.

There were multiple issues with using tags:

- Running unit tests would filter out IntegrationTest tagged tests, meaning it would generate an empty report for all integration test suites. This led to further problems with us needing to explicitly output test results to different folders.
- It was unclear how to run a single integration test; testOnly commands did not work as expected.
- The sbt configuration was difficult to understand