Fix Mesos HTTP health check for non-host networking mode with `containerPort=0` (#6235)
Fixed a bug where for a container in non-host networking mode (`container` and `container/bridged`) with `hostPort=0, containerPort=0` we would set the `containerPort == hostPort` in `ContainerInfo` *and* ports environment variables *but not* in the `NetworkInfo`.
As the result Mesos health checks in this scenario would try to probe `containerPort == 0`. This behavior has now been unified for `NetworkInfo`, `ContainerInfo` and generated ports environment variables resulting in fixed Mesos HTTP health checks.