Schedule and then provision instance. (#6181)
Summary:
Currently we create an instance when we accept an offer. This makes
tracking hard. Furthermore the `TaskLauncherActor` has a simple count
instead of a list of instances to launch. This change creates instances
in a scheduled state before there is any offer matching. The
`TaskLauncherActor` fetches a new list of instances form the
`InstanceTracker` and matches offers for scheduled instances. Once there
is a match the instance is updated into a provisioned state.

We coming updates we probably want to push the instance map with an
offer instead of keeping a copy in the `TaskLauncherActor`. The
`LaunchQueueActor` can go since a scheudle is a simple create operation.

JIRA issues: MARATHON-8169, MARATHON-8170, MARATHON-8171