Non-blocking IO Leader Proxy (#6193)
We modify the proxy logic to use non-blocking, asynchronous IO.

For reading Jetty servlet posted entities, and writing Jetty servlet responses, we use the non-blocking async reading / writing method as described here:

    https://docs.oracle.com/javaee/7/tutorial/servlets012.htm

Further, we switch from the native Java HTTP client to Akka HTTP. Unfortunately, in so doing, we lost the ability to distinguish between TCP timed-out connections, and connection refused. This is a subtle difference which we do not know if it will be meaningful. A follow up PR to enable user to switch back to synchronous logic will follow.

JIRA Issue: MARATHON-8174