Remove unreachable resident task terminal status surrogate
Before Mesos 1.4.0, on reboot agents would come up with different IDs. This
meant that unreachable tasks on the node would never become terminal, forever
blocking the recovery of resident tasks. In MARATHON-2311 we created a
mechanism to treat an offer with reservations for a task as an indication of
the task being gone.

That is fixed in Mesos 1.4.0 and I have confirmed that terminal statuses are
now received for unreachable tasks when rebooted hosts come back online.

JIRA Issues: MARATHON-8272