Pull in shakedown as an editable dependency.
Summary:
We have a lot of Shakedown changes pending in `tests/system/common.py`.
So far it takes a few steps to move the changes to Shakedown and then
pull them back again. This patch eases the process by importing
Shakedown as a git-subrepo. It alos to make changes to Shakedown in the
Marathon repository *and* push changes upstream to create pull requests
for Shakedown. This way Shakedown in Marathon will be slightly ahead of
the Shakedown master *but* also ensure that the Shakedown master will
stay uptodate.