Import performance test Jenkins job definition. (#6310)
* Import performance test Jenkins job definition.

Summary:
The current Jenkins job definition that runs performance tests against master
and reports it to DataDog is not version controlled. This makes
iterating over the tests hard. This diff defines yet another Jenkinfiles
and imports the bash script verbatim. Once `ic/perf/5k-with-timeseries`
is merged into master we can define a Jenkins job or pipeline using the
definition of this diff and start iterating.

* Remove AWS credentials.

* Remove DataDog app key.

* Remove DataDog APP key parameter.