Move the instance update to the update method (#6335)
Summary:
I think the reservation update should happen in the instance update method. Someone in a different PR commented on this piece of code in a way that it seemed like it might be misleading so I refactored it. It is supposed to work the same as it worked before :)