Remove unreachable resident task terminal status surrogate (#6290)
Before Mesos 1.4.0, on reboot agents would come up with different IDs. This
meant that unreachable tasks on the node would never become terminal, forever
blocking the recovery of resident tasks. In MARATHON-2311 we created a
mechanism to treat an offer with reservations for a task as an indication of
the task being gone.

That is fixed in Mesos 1.4.0 and I have confirmed that terminal statuses are
now received for unreachable tasks when rebooted hosts come back online.

Additionally, a needed operator call was introduced in Mesos 1.5.0 to enable
users to mark agents as gone when their metadata is partially wiped such that
the agent ID changes but reservations stay the same. As such, we set the
minimum Mesos version to 1.5.0.

We also update the docs.

JIRA Issues: MARATHON-8272