Crash if ZK DNS is not resolvable (#6412)
Currently, if Marathon can't connect to ZK because it's name is resolvable, Guice will reinstantiate `StorageModule` N times, effectively retrying to connect to ZK N times. Marathon will eventually exit with exitCode=137 but it takes ~5min with default settings. With this change we crash on the first error.

JIRA Issues: MARATHON-7390