Unify `deployment_wait` in the SI tests (#6440)
Summary:
Currently, we're using `deployment_wait` waiting for all existing deployments to finish. This makes our SI tests more brittle and dependent on each other. This change replaces that with waiting for an app/pod id specific to the test in most cases. Besides that, we're now using `assert_that` matcher in `common.deployment_wait`