Import the dcos-cli python library.
Summary:
This patch imports the source code of the dcos-cli python library from
tag 0.5.7 (dcos/dcos-cli@1e8d845). The CLI team moved over to a Go based
client and dropped support for the Python client. We only require a few
tools. This is the first step for refactoring.

JIRA issues: MARATHON-8384