Create GitHub pull requests statistics.
Summary:
We want to move faster and reduce the number of our pull requests. Our
first step is to generate data and get an overview. The Python script
`ci/github_pulls_stats.py` gets the list of pull requests and calculates
the youngest, oldest, p50 and p90 open pull request.