Remove cli part of Shakedown and import credentials.
Summary:
This diff remove the last CLI part from Shakedown. It does not wrap
pytest anymore but proveds a collection of tools and clients for
testing.

The DC/OS url, SSH user and key and cluster credentials should be passed
with environment variables. These are backward compatible with the
Shakedown CLI environment variables.

JIRA issues: MARATHON-8385