Address case where we sometimes eliminate an extra Standby(None) during leader transition events (#6525)
LeaderLatch.getParticipants will return records with no current leader, if the ephemeral leader record is deleted
between the getChildren operation and the getData call for the record in question. This means that sometimes, we can go from:

- LeadershipState.Standby(leader1)
- LeadershipState.Standby(None)
- LeadershipState.ElectedAsLeader

In the real world, this extra event is harmless, but having completely
deterministic tests helps us detect if there are edge cases we may
have missed.