Bump Jekyll dependencies
Several Jekyll dependencies were out of date and this was causing known vulnerability messages to show in GitHub. Bumping this resolves it.

Jekyll was having issues with UTF-8 emdashes in our css files; we updated the Docker image to handle them.

Further, we clarify the docs and restore a section for simply rendering the docs on-the-fly.