Use one queue to process LaunchQueue.Add commands.
Summary:
The LaunchQueueActor will go with MARATHON-8167. However, until then we
are stuck with it. Due to its design there could be a case of concurrent
updates of its state. We could solve this either by introducing more
messages or with an internal queue. We opted for the second solution and
intorduced a queue that gurantess sequential processing. It's ugly and
this diff just simplifies it to one queue. This should avoid errors in
the processing. Eg the current multi queue solution could stall and
ignore queued items.