Launch instance based on scheduled state. (#6340)
Summary:
We would like to create instances in a scheduled state *before* they are
launched so that we can avoid race conditions.

That is why we introduce a `Scheduled` and `Provisioned` condition.

The instance tracker will have the ground truth to scheduled instances.
The task start actor pulls a list of the instances and matches offers
against the instances that are considered scheduled.

All logic in the orchestration layer, eg `TaskReplaceActor`, only
queries and sets the goal of a instance instead of the current state.
This avoids race conditions as the state might be change by different
entities. The goal is only change in one orchestrator at a time.

Since the instance tracker has the ground truth for scheduled instance
the launch queue should not be queried any more but only the tracker. It
is a single point.

For newly scheduled residents the flow is following:
```
Scheduled -> Scheduled (with Reservation) -> Provisioned -> ...
```

For relaunching residents the flow is:
```
Reserved (terminal) -> Scheduled (with Reservation) -> Provisioned -> ...
```

Scheduled instances with reservations have agentinfo now. That's necessary so
we don't violate placement constraints and e.g. do reservation on agent that
already have one scheduled and has hostname: unique.


Individual commits:

* Introduce `Provisioned` and `Scheduled` condition. (#6180)
* Schedule and then provision instance. (#6181) 
* Remove launch queue querying in favour of instance tracker. (#6233)
* Resident task should be scheduled in the same way as other tasks (#6264)
* Launch inactive instances with goal Running. (#6344)
* Stabilization fixes for resident tasks (#6373)
* Test that Stopped instance is not expunged (#6371)
* Remove RescheduleReserved as it is not used anymore
* Use rescheduleReserved - we need it to change the condition from Reserved (#6378)
* Do not process multiple `Launch.Add` in parallel for the same runSpec (#6376)
* Do not use outdated state from TaskReplaceActor
* Improve the log message and docs (#6380)
* Properly start TaskLauncherActor after failover (#6394)
* Reschedule instance with provision timeout. (#6418)
* Use one queue to process LaunchQueue.Add commands. (#6534)

JIRA issues: MARATHON-8167

Co-authored-by: Alena Varkockova <varkockova.a@gmail.com>