`TaskOverdueActor` now respects `--kill_retry_timeout` parameter
Summary:
previously `TaskOverdueActor` would periodically check for overdue tasks and request killing them which would lead to too
many kill requests being send to Mesos due to a bug in the `KillService`. This is now fixed. In detail:
- `KillService` does not send kill requests to Mesos for tasks that are already "in flight"
- a new `KillService.killInstancesAndForget()` method was introduced for cases where the caller doesn't care about the
result. In this case `KillStreamWatcher` will not be created potentially preventing a memory leak (should there be stuck
tasks for a long period of time in the cluster)
- `kill_retry_timeout` was increased to 30 seconds by default. This should further prevent overloading agents with kill
requests (actually, it's the docker daemon who is easily overloaded but anyway)

Fixes: MARATHON-8453