Write out correct version in ZkId. (#6601)
Summary:
With #6558 we started to include the micro seconds in app versions even
when they were zero. Older Marathon versions would drop the zeros. We
would not find apps from previous versions though since the string representation is different.

Commit that introduced storing timestamps with trailing zeros was reverted (since it generated a lot of flakes), and now we truncate microsecond instead of changing the formatter.

JIRA issues: MARATHON-8461

Co-authored-by: Nikita Melkozerov nmelkozerov@mesosphere.io