Fix issue in which Stats failed to initialize
There's an ugly order-of-instantiation issue for our WhenLeaderActor
instances. Ultimately it seemed easiest to implement an annotation to
allow certain messages to be delayed until an actor is leader, rather
than create a complex, probably racy order-of-initialization scheme,
or instantiate the stats module lazilly (also probably racy... and
awful. History is litered with flaky non-deterministic boot sequences)