Remove `InstanceUpdateActor`. (#6598)
Summary:
The async call into the `InstanceUpdateActor` is not required anymore
since the instance tracker can pass the instances. We used to have it async
because the updater would call into the repository. That changed with #6177.

The updates for a single instance a sequentialized in an Akka stream in
`InstanceTrackerDelegate`.

Co-authored-by: Aleksey Dukhovniy <adukhovniy@mesosphere.io>

JIRA Issues: MARATHON-8470