Replaced `AppStartActor` with a stub (#6678)
Summary:
Previously `AppStartActor` was responsible for handling `StartApplication` deployment step. However, it's logic was completely duplicating `ScaleApplication` deployment step. This change stubs the `AppStartActor` with a future that is immediately successfully completed. The reason, it is still present in the code base is because it is part of our public Event API and someone might rely on it.

Additionally, the `SchedulerActions` object that was part of the deployment actor got removed from it since it's not used anymore. By now, it should be only used by the periodic scaling check and will hopefully become obsolete soon.