Do not put whole instance into provision update op (#6694)
Summary:
InstanceUpdateOperation Provision contained the whole instance. Instance update operations might be queued so it might happen that once the update is applied, that instance we enqueued is already obsolete and the instance changed (e.g. the goal changed).

With this change, we always work with the newest instance while applying provision update.

JIRA issues: MARATHON-8516