Stable instance ids (#6677)
Summary:
Until now, Marathon expunged instances once all of their tasks ended up in a terminal state, and eventually launched replacements as a result of those instances being expunged.
This commit adjusts this behavior. Instances will now only be expunged from the state when their goal is to be Decommissioned and all their tasks are in a terminal state. If their goal is still Running, they will be considered to be scheduled and used to launch replacement tasks.

This means that instance Ids are now stable for as long as an instance shall be kept running. New instances will be created only when replacing unreachable instances, and when replacing instances with new versions. Similar to the way we handle task Ids for stateful services, tasks of stateless services will now also provide an incarnation count, appended to the task Id. The first task created for an instance will be the .1, and subsequent replacements will increment that incarnation counter.
When killing an instance using the wipe=true flag, its goal will be set to decommissioned, and it will eventually be expunged when all tasks are terminal. Note that as long as its tasks are e.g. unreachable, it will not be expunged until they are reported terminal (GONE, GONE_BY_OPERATOR, or UNKNOWN).

When killing instances without the wipe=true or scale=true flags, Marathon will only issue kill requests to Mesos, but keep the current goal and will, therefore, launch replacements that are still associated with the existing instance.

JIRA Issues: MARATHON-8373, MARATHON-8140

Co-authored-by: Matthias Eichstedt <matthias.eichstedt@mesosphere.io>
Co-authored-by: Tim Harper <timcharper@gmail.com>
Co-authored-by: Karsten Jeschkies <k@jeschkies.xyz>