Attach run spec to instance. (#6661)
Summary:
Each instance hold a run spec id and the run spec version. This
is a unique identifier for the run spec. Thus we could just attach
the run spec itself to the instance. This will simplify our launching
and matching logic. It will also open the door to prototype templates.

See #6664 for an illustration. 

JIRA issues: MARATHON-8325