Introduce reservation id type. (#6687)
Summary:
The new reservation id type makes the difference between a task id,
instance id and reservation id clear. So far the are very similar and it
is trivial to construct one from the other. However, in the near future
we would like to reassign a reservation to a different instance. That
means reservation and instance ids have to be independent. This is the
first step.

Depends on #6681.