Re-introduce jenkins `timeout` for SI tests
We already use `timeout` for the SI test execution, however it only covers the tests themselves and not cluster creation/removal. This leads to some "stuck jobs" in cases where e.g. dcos-launch is stuck removing the cluster. This is hardly a proper solution but at least we don't overload jenkins with never-ending jobs.