Extract ir and backend from type generator monolith. (#6743)
Summary:
This is the first step to disentangle the RAML type generator code. As
once can see the backend, ie Treehugger code generation, leaks into the
IR, ie type definitions. We can solve this in another step with a
visitor pattern. However, until we have proper tests we can just move
code as in this diff.

Some IR is left in `RamlTypeGenerator`.