Use InstanceTracker subscription mechanism to wait for instance terminality
This approach race condition free; one does not need to
register the KillStreamWatcher before the kill requests are sent, as
kills will be detected if they'd already happened before the watcher
was registered. This is achieved by recognizing one the following states
as "terminal"

- Instance was removed from the instance tracker
- Instance has running tasks with Ids of later incarnation
- Instance is considered terminal (unreachable, killed, etc.)

Further, a mechanism was introduced to wait for instances to become
decomissioned. This simply waits for a moment in time where the
instance is no longer in the instanceTracker, either during the
initial snapshot or in subsequent updates.