Add `?containerNetworks` querystring parameter to /v2/tasks text output
Backport of #7123

In this commit we fixed the plaintext output of /v2/tasks to no longer
include nonsensical addresses for container endpoints that lacked a host
port mapping. Further, we introduce the parameter "containerNetworks" to
opt in to include non-host mapped container endpoints.

containerNetworks is a comma delimited list of network names, where the
value "*" is interpreted as including all container networks.

JIRA issues: MARATHON-8721